Update documentation for running Rust solutions
Added the `--release` flag to the cargo command for running solutions. This
builds the executables with optimizations and without debug symbols. For full
details on what the `release` profile does see Cargo's documentation:
https://doc.rust-lang.org/cargo/reference/profiles.html#release.

Users running solutions that are tested and known to be working properly do not
need the `debug` profile. Use the `release` profile should speed up running the
solutions (sometimes by a lot).